+++ core/src/main/kotlin/org/pastalab/fray/core/concurrency/LockHoldingInfo.kt
@@ -1,0 +1,3 @@
+package org.pastalab.fray.core.concurrency
+
+data class LockHoldingInfo(val resourceId: Long) {}