+++ core/src/main/kotlin/org/pastalab/fray/core/concurrency/operations/CountDownLatchWaitOperation.kt
@@ -1,10 +1,11 @@
 package org.pastalab.fray.core.concurrency.operations
 
 import org.pastalab.fray.core.concurrency.primitives.CountDownLatchContext
-import org.pastalab.fray.core.concurrency.primitives.InterruptionType
+import org.pastalab.fray.rmi.ResourceInfo
+import org.pastalab.fray.rmi.ResourceType
 
 class CountDownLatchAwaitBlocking(timed: Boolean, val latchContext: CountDownLatchContext) :
-    TimedBlockingOperation(timed) {
+    BlockedOperation(timed, ResourceInfo(latchContext.latchId, ResourceType.CDL)) {
   override fun unblockThread(tid: Long, type: InterruptionType): Any? {
     if (latchContext.unblockThread(tid, type)) {
       return this