+++ core/src/main/kotlin/org/pastalab/fray/core/concurrency/operations/TimedBlockingOperation.kt
@@ -1,5 +1,0 @@
-package org.pastalab.fray.core.concurrency.operations
-
-import org.pastalab.fray.core.concurrency.primitives.Interruptible
-
-abstract class TimedBlockingOperation(val timed: Boolean) : NonRacingOperation(), Interruptible {}