+++ core/src/main/kotlin/org/pastalab/fray/core/concurrency/primitives/Acquirable.kt
@@ -1,0 +1,5 @@
+package org.pastalab.fray.core.concurrency.primitives
+
+import org.pastalab.fray.rmi.ResourceInfo
+
+abstract class Acquirable(val resourceInfo: ResourceInfo)