+++ core/src/main/kotlin/org/pastalab/fray/core/concurrency/primitives/InterruptibleContext.kt
@@ -1,5 +1,7 @@
 package org.pastalab.fray.core.concurrency.primitives
 
+import org.pastalab.fray.core.concurrency.operations.InterruptionType
+
 interface InterruptibleContext {
   /** A primitive should implement this method if it supports interruption. */
   fun unblockThread(tid: Long, type: InterruptionType): Boolean