+++ rmi/src/main/kotlin/org/pastalab/fray/rmi/RemoteScheduler.kt
@@ -13,11 +13,24 @@ enum class ThreadState {
   Created,
 }
 
+enum class ResourceType {
+  LOCK,
+  CONDITION,
+  SEMAPHORE,
+  CDL,
+  PARK,
+  SYNCURITY_CONDITION,
+}
+
+data class ResourceInfo(val resourceId: Int, val resourceType: ResourceType)
+
 data class ThreadInfo(
     val threadName: String,
     val index: Long,
     val state: ThreadState,
     val stackTraces: List<StackTraceElement>,
+    val waitingOn: ResourceInfo?,
+    val acquired: Set<ResourceInfo>
 ) : Serializable
 
 interface RemoteScheduler : Remote {